backend.flask/pyproject.toml
[tool.poetry]
name = "backend-flask"
version = "0.1.0"
description = ""
authors = ["P4SSER8Y <P4SSER8Y@outlook.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
gunicorn = "^20.1.0"
loguru = "^0.6.0"
pyyaml = "^6.0"

[[tool.poetry.source]]
name = "tencent"
url = "http://mirrors.cloud.tencent.com/pypi/simple"
default = true
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
